fastapi_backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai-whisper
torch
numpy==1.26.4
